releaseherald/requirements.txt
boltons
click
GitPython
Jinja2
toml
pydantic
pluggy
typing_extensions